==> PANDAS/2-dataframe.ipynb <==
import pandas as pd

dados = {
    'Nome': ['Alice', 'Bob', 'Charlie'],
    'Idade': [24, 30, 22],
    'Cidade': ['São Paulo', 'Rio de Janeiro', 'Belo Horizonte']
}

df = pd.DataFrame(dados)
df

# Head traz por padrão as 5 primeiras linhas, mas recebe um inteiro como parametro
df.head(1)

# Tail traz por padrão as 5 ultimas linhas, mas recebe um inteiro como parametro
df.tail(1)

dados_notas = {
    'Nome': ['Alice', 'Bob', 'Charlie', 'David', 'Eva'],
    'Matemática': [85, 90, 78, 88, 92],
    'Física': [80, 75, 88, 85, 95],
    'Química': [78, 82, 85, 80, 88]
}

df_notas = pd.DataFrame(dados_notas)
df_notas

# Incluir coluna Média no dataframe e calcular média das notas por aluno

df_notas['Média'] = df_notas[['Matemática','Física', 'Química']].mean(axis=1)
df_notas

# Identificar alunos aprovados com nota boa (média >= 82.5)
alunos_aprovados = df_notas[df_notas['Média'] >= 82.5]
alunos_aprovados

# Exibir a Média e os alunos aprovados

df_notas[['Nome', 'Média']]

alunos_aprovados[['Nome', 'Média']]
