==> PANDAS/5-importa_csv.ipynb <==
import pandas as  pd

Importar arquivo CSV

df_produtos = pd.read_csv("dados/produtos.csv")
df_produtos

df_produtos.head(10)

df_produtos.tail(8)

Shape -> Mostra o formato da Matriz

df_produtos.shape

dtypes -> Mostra os tipos de dados das colunas

df_produtos.dtypes

Ordenando por quantidade '(sort_values(by='Quantidade', ascending=False))'
- by='Coluna'
- ascending= True -> ordenação decrescente
- ascending= False -> ordenação decrescente

df_produtos.sort_values(by='Quantidade', ascending=False)

Agrupar por determinada colunas e saber quanto itens está relacionado
- dataframe['Coluna'].value_counts()


df_produtos['Categoria'].value_counts()
