==> PANDAS/6-analise_csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv('dados/produtos.csv')
df_produtos

df_produtos.dtypes

Script para transformar uma coluna do tipo numerico

df_produtos['Quantidade'] = pd.to_numeric(df_produtos['Quantidade'])
df_produtos['Preço'] = pd.to_numeric(df_produtos['Preço'])

Desenpenho de Vendas por Categoria

vendas_categoria = df_produtos.groupby('Categoria')['Quantidade'].sum()
vendas_categoria

Produto mais vendido

produto_mais_vendido = df_produtos.groupby('Produto')['Quantidade'].sum().idxmax()
produto_mais_vendido

Total de unidades Vendidas por Produtos

produto_mais_vendido = df_produtos.groupby('Produto')['Quantidade'].sum()
produto_mais_vendido

Preço Médio por Produto

media_preco_produto = df_produtos.groupby('Produto')['Preço'].mean()
media_preco_produto

Correlação entre Unidades Vendidas e Preço Unitário

correlacao = df_produtos['Quantidade'].corr(df_produtos['Preço'])
correlacao

Desenpenho por Vendas x Categoria -> Gerando novo Index e novos nomes para as Colunas

vendas_categoria_df = vendas_categoria.reset_index()
vendas_categoria_df.columns = ['Categoria', 'UnidadesVendidas']
vendas_categoria_df

Exportando o arquivo para um CSV

vendas_categoria_df.to_csv('dados/exportado_na_aula_6.csv', index=False)
