==> PANDAS/7-criar_planilha.ipynb <==
import pandas as pd

dados = {
    'Times': ['Flamengo', 'Palmeiras', 'São Paulo', 'Santos', 'Corinthians'],
    'Partidas_Jogadas': [10, 10, 10, 10, 10],
    'Vitorias': [7, 5, 6, 4, 3],
    'Empates': [2, 3, 1, 3, 2],
    'Derrotas': [1, 2, 3, 3, 5],
    'Gols_Pro': [22, 18, 16, 14, 10],
    'Gols_Contra': [10, 12, 15, 16, 20]
}

df = pd.DataFrame(dados)
df

Gerando nova coluna 'Saldo_Gols'
- Gols_Pro * Gols_Contra

df['Saldo_Gols'] = df['Gols_Pro'] - df['Gols_Contra']
df

Gerando nova coluna 'Pontos'
- Vitorias * 3 + Empates

df['Pontos'] = (df['Vitorias'] * 3) + (df['Empates'])
df

Ordenando a lista por pontos

df = df.sort_values(by='Pontos', ascending=False)
df

Exportando o arquivo para um planilha em Excel

df.to_excel('dados/campeonato_futebol.xlsx', index=False)
