==> PANDAS/8-analise_planilha.ipynb <==
import pandas as pd

df = pd.read_excel('dados/campeonato_futebol.xlsx')
df

df.dtypes

df.describe()

df.head(2)

df.tail(2)

Analise de desempenho por time

- Especificando as colunas desejadas [['Colunas_entre_parenteses_suplos']]
- Setando index na coluna 'Times' --> set_index('Times')

desempenho_info = df[['Times', 'Pontos', 'Vitorias', 'Saldo_Gols', 'Gols_Pro', 'Gols_Contra']].set_index('Times')
desempenho_info

- Times com mais Vitórias
- Times com menos Vitórias

time_mais_vitoria = df[df['Vitorias'] == df['Vitorias'].max()]
time_mais_vitoria


time_menos_vitorias = df[df['Vitorias'] == df['Vitorias'].min()]
time_menos_vitorias

Distribuição por Gols a Favor (Gols_Pro)

destribuicao = df.groupby('Times')['Gols_Pro'].value_counts()
destribuicao

Verificar a correlação entre Gols_Pro x Ponto
- Quem mais fez Gols/Melhor ataque, mais fez ponto?

correlacao = df['Gols_Pro'].corr(df['Pontos'])
correlacao

correlacao = df['Gols_Contra'].corr(df['Pontos'])
correlacao
