==> PANDAS/10-importa_json.ipynb <==
import pandas as pd

df = pd.read_json('dados/receitas.json', encoding='utf-8')
df

df_receitas = pd.json_normalize(
    df['receitas'],
    'ingredientes',
    ['titulo', 'tempo_preparo', 'porcoes']
)

df_receitas
