==> PANDAS/12-bd.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

engine = create_engine("sqlite:///ecommerce.db")

df = pd.read_csv('dados/produtos.csv')
df

df.head()

try:
    df.to_sql('produtos', con=engine, if_exists='replace', index=False)
    print("Dados exportados com sucesso para a tabela 'produtos'")
except Exception as e:
    print(f'ocorreu um erro ao exportar os dados: {e}')
