==> PANDAS/13-analise_bd.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

Conexão com banco de dados

engine = create_engine("sqlite:///ecommerce.db")

Primeira consulta no banco de dados com Pandas
- Total de Vendas por Categoria

consulta_1 = "SELECT categoria, SUM(Quantidade) AS Total_Qtd_Vendida FROM produtos GROUP BY categoria"
resultado_1 = pd.read_sql_query(consulta_1, engine)
resultado_1

Produto com mais vendas

consulta_2 = "SELECT Produto, SUM(Quantidade) AS Total_Qtd_Vendida FROM produtos GROUP BY Produto ORDER BY Total_Qtd_Vendida DESC LIMIT 1"
resultado_2 = pd.read_sql_query(consulta_2, engine)
resultado_2

Total de Vendas por Produto

consulta_3 = "SELECT Produto, SUM(Quantidade) AS Total_Qtd_Vendida FROM produtos GROUP BY Produto ORDER BY Total_Qtd_Vendida DESC"
resultado_3 = pd.read_sql_query(consulta_3, engine)
resultado_3

Média de Preço por Produto

consulta_4 = "SELECT Produto, AVG(Preço) AS Media_Preço FROM produtos GROUP BY Produto"
resultado_4 = pd.read_sql_query(consulta_4, engine)
resultado_4
